refactor: ScanDialog 解耦重构、自适应视图还原、数据显示三模式分流过滤 (警告修复版)
1. 全局解耦：剥离无边框窗口拉伸/拖动至独立的 ResizeEventFilter 全局拦截器；将 ScanConfig 隔离到 ConfigManager 单例中；引擎生命周期在应用退出时安全清理。
2. 模块化：引入 IScanResultView 接口及 ListResultView、JustifiedResultView 和 GridResultView 三大排版视图子模块，使用视图适配器模式（View Adaptor Pattern）实现 ScanDialog 的极简多态代理。
3. 行为还原：将 JustifiedMode (自适应模式) 布局算法彻底还原为“旧版本-1”的原生自适应流式排布形态。
4. 模式感知过滤：在 ScanTableModel::updateResults() 中引入模式感知过滤。自适应与网格模式作为媒体画廊，在源头过滤仅保留视频与图像后缀文件；而列表模式继续展示全量搜索数据。
5. 警告消除：使用 static_cast<int>(i) 强转消除了 64位编译下的 C4267 数据截断警告。
6. 编译修复：完美解决了之前版本由于 IScanResultView 构造函数参数类型不匹配而导致的编译阻碍。

+++ src/ui/ScanDialog.cpp
@@ -671,7 +671,7 @@ void ScanTableModel::updateResults(std::shared_ptr<ResultSet> nextSet) {
         // 重建过滤后结果集的 O(1) 反向索引映射
         newSet->keyToPos.clear();
         for (size_t i = 0; i < newSet->keys.size(); ++i) {
-            newSet->keyToPos[newSet->keys[i]] = i;
+            newSet->keyToPos[newSet->keys[i]] = static_cast<int>(i);
         }
     } else {
         // 列表模式：保留全量普通文件、文件夹及多媒体过滤数据
